Move core DoodleBot functionality to separate file/class
